kill 前判断是否是假人

+++ src/main/java/dev/dubhe/gugle/carpet/commands/BotCommand.java
@@ -216,6 +216,7 @@ public class BotCommand {
         for (String botName : botNames) {
             ServerPlayer player = source.getServer().getPlayerList().getPlayerByName(botName);
             if (player == null) continue;
+            if (!(player instanceof EntityPlayerMPFake)) continue;
             player.kill();
         }
         return 1;
